first commit refact(change)

+++ src/main/java/credit/CreditCalculator.java
@@ -3,15 +3,16 @@ package credit;
 
 public class CreditCalculator {
 
-    public int monthPaymentCalc(int sum, int percent, int months) {
+
+    public double monthPaymentCalc(int sum, int percent, int months) {
         return 0;
     }
 
-    public int totalAmount(int sum, int percent, int months) {
+    public double totalAmount(int sum, int percent, int months) {
         return 0;
     }
 
-    public int allOverpayments(int sum, int percent, int months) {
+    public double allOverpayments(int sum, int percent, int months) {
         return 0;
     }
 }
